UserService_Flask/requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
SQLAlchemy==2.0.30
pymssql==2.3.0
python-dotenv==1.0.1